+++ src/main/scala/com/ardentex/spark/hiveudf/FormatTimestamp.scala
@@ -12,20 +12,20 @@ import org.apache.hadoop.hive.ql.exec.UDF
 class FormatTimestamp extends UDF {
 
   def evaluate(t: Timestamp, fmt: String): String = {
-    if ((t == null) || (fmt == null)) {
-      ""
-    }
-    else {
-      try {
+    val optRes =
+      for { ts <- Option(t)     // null check
+            f  <- Option(fmt) } // null check
+      yield try {
         val formatter = new SimpleDateFormat(fmt)
         formatter.format(new Date(t.getTime))
       }
       catch {
-        // Bad format possibly. Return Timestmap.toString. (We could return
+        // Bad format. Return Timestmap.toString. (We could return
         // an error message, as well, but this is fine for now.)
-        case _: Exception =>
+        case _: IllegalArgumentException =>
           t.toString
       }
-    }
+
+    optRes.getOrElse("")
   }
 }