+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 0.1.0
 
 - Updated to compile against both Scala 2.10 and 2.11.
 - Verified against Spark 2.1.
+- Made Scala UDFs more idiomatic.
 
 Version 0.0.1